requirements.txt
fire~=0.3.1
PyGithub~=1.53
PyYAML~=5.3.1
